refactor: Consolidate pickup-history endpoint route (#108)
Move GetPickupHistory from PickupController to AuthorizedPickupController
to fix inconsistent API route paths.

Route changes:
- Before: /api/v1/checkin/people/{childIdKey}/pickup-history
- After: /api/v1/people/{childIdKey}/pickup-history

This groups all people-related pickup endpoints under /api/v1/people:
- GET /api/v1/people/{childIdKey}/authorized-pickups
- GET /api/v1/people/{childIdKey}/pickup-history

Additional improvements:
- RecordPickup now returns proper 201 Created with Location header
- Added ChildIdKey to PickupLogDto for Location header construction
- Created AuthorizedPickupControllerTests with 4 comprehensive tests
- Updated PickupControllerTests for new return type
