+++ tests/Koinon.Api.Tests/Controllers/PickupControllerTests.cs
@@ -424,6 +424,7 @@ public class PickupControllerTests
         var expectedLog = new PickupLogDto(
             IdKey: IdKeyHelper.Encode(300),
             AttendanceIdKey: _attendanceIdKey,
+            ChildIdKey: _childIdKey,
             ChildName: "Johnny Smith",
             PickupPersonName: "Sarah Smith",
             WasAuthorized: true,
@@ -440,7 +441,9 @@ public class PickupControllerTests
         var result = await _controller.RecordPickup(request);
 
         // Assert
-        var createdResult = result.Should().BeOfType<CreatedAtActionResult>().Subject;
+        var createdResult = result.Should().BeOfType<CreatedResult>().Subject;
+        createdResult.StatusCode.Should().Be(StatusCodes.Status201Created);
+        createdResult.Location.Should().Be($"/api/v1/people/{_childIdKey}/pickup-history");
         var pickupLog = createdResult.Value.Should().BeAssignableTo<PickupLogDto>().Subject;
         pickupLog.WasAuthorized.Should().BeTrue();
         pickupLog.SupervisorOverride.Should().BeFalse();
@@ -489,6 +492,7 @@ public class PickupControllerTests
         var expectedLog = new PickupLogDto(
             IdKey: IdKeyHelper.Encode(300),
             AttendanceIdKey: _attendanceIdKey,
+            ChildIdKey: _childIdKey,
             ChildName: "Johnny Smith",
             PickupPersonName: "Unknown Person",
             WasAuthorized: false,
@@ -505,7 +509,9 @@ public class PickupControllerTests
         var result = await _supervisorController.RecordPickup(request);
 
         // Assert
-        var createdResult = result.Should().BeOfType<CreatedAtActionResult>().Subject;
+        var createdResult = result.Should().BeOfType<CreatedResult>().Subject;
+        createdResult.StatusCode.Should().Be(StatusCodes.Status201Created);
+        createdResult.Location.Should().Be($"/api/v1/people/{_childIdKey}/pickup-history");
         var pickupLog = createdResult.Value.Should().BeAssignableTo<PickupLogDto>().Subject;
         pickupLog.SupervisorOverride.Should().BeTrue();
         pickupLog.SupervisorName.Should().Be("John Supervisor");
@@ -534,136 +540,6 @@ public class PickupControllerTests
         var problemDetails = badRequestResult.Value.Should().BeOfType<ProblemDetails>().Subject;
         problemDetails.Title.Should().Be("Invalid request");
         problemDetails.Detail.Should().Contain("AttendanceIdKey is required");
-    }
-
-    #endregion
-
-    #region GetPickupHistory Tests
-
-    [Fact]
-    public async Task GetPickupHistory_WithValidChildIdKey_ReturnsOkWithHistory()
-    {
-        // Arrange
-        var expectedHistory = new List<PickupLogDto>
-        {
-            new(
-                IdKey: IdKeyHelper.Encode(300),
-                AttendanceIdKey: _attendanceIdKey,
-                ChildName: "Johnny Smith",
-                PickupPersonName: "Sarah Smith",
-                WasAuthorized: true,
-                SupervisorOverride: false,
-                SupervisorName: null,
-                CheckoutDateTime: DateTime.UtcNow.AddDays(-1),
-                Notes: null),
-            new(
-                IdKey: IdKeyHelper.Encode(301),
-                AttendanceIdKey: IdKeyHelper.Encode(101),
-                ChildName: "Johnny Smith",
-                PickupPersonName: "Mike Smith",
-                WasAuthorized: true,
-                SupervisorOverride: false,
-                SupervisorName: null,
-                CheckoutDateTime: DateTime.UtcNow.AddDays(-7),
-                Notes: null)
-        };
-
-        _authorizedPickupServiceMock
-            .Setup(s => s.GetPickupHistoryAsync(
-                _childIdKey,
-                null,
-                null,
-                It.IsAny<CancellationToken>()))
-            .ReturnsAsync(expectedHistory);
-
-        // Act
-        var result = await _supervisorController.GetPickupHistory(_childIdKey, null, null);
-
-        // Assert
-        var okResult = result.Should().BeOfType<OkObjectResult>().Subject;
-        var history = okResult.Value.Should().BeAssignableTo<List<PickupLogDto>>().Subject;
-        history.Should().HaveCount(2);
-        history[0].ChildName.Should().Be("Johnny Smith");
-        history[1].PickupPersonName.Should().Be("Mike Smith");
-    }
-
-    [Fact]
-    public async Task GetPickupHistory_WithInvalidDateRange_ReturnsBadRequest()
-    {
-        // Arrange
-        var fromDate = new DateTime(2024, 1, 15);
-        var toDate = new DateTime(2024, 1, 10); // toDate before fromDate
-
-        // Act
-        var result = await _supervisorController.GetPickupHistory(_childIdKey, fromDate, toDate);
-
-        // Assert
-        var badRequestResult = result.Should().BeOfType<BadRequestObjectResult>().Subject;
-        var problemDetails = badRequestResult.Value.Should().BeOfType<ProblemDetails>().Subject;
-        problemDetails.Title.Should().Be("Invalid date range");
-        problemDetails.Detail.Should().Contain("toDate' must be greater than or equal to 'fromDate");
-    }
-
-    [Fact]
-    public async Task GetPickupHistory_WithDateRange_ReturnsFilteredHistory()
-    {
-        // Arrange
-        var fromDate = new DateTime(2024, 1, 1);
-        var toDate = new DateTime(2024, 1, 31);
-
-        var expectedHistory = new List<PickupLogDto>
-        {
-            new(
-                IdKey: IdKeyHelper.Encode(300),
-                AttendanceIdKey: _attendanceIdKey,
-                ChildName: "Johnny Smith",
-                PickupPersonName: "Sarah Smith",
-                WasAuthorized: true,
-                SupervisorOverride: false,
-                SupervisorName: null,
-                CheckoutDateTime: new DateTime(2024, 1, 15, 12, 0, 0),
-                Notes: null)
-        };
-
-        _authorizedPickupServiceMock
-            .Setup(s => s.GetPickupHistoryAsync(
-                _childIdKey,
-                fromDate,
-                toDate,
-                It.IsAny<CancellationToken>()))
-            .ReturnsAsync(expectedHistory);
-
-        // Act
-        var result = await _supervisorController.GetPickupHistory(_childIdKey, fromDate, toDate);
-
-        // Assert
-        var okResult = result.Should().BeOfType<OkObjectResult>().Subject;
-        var history = okResult.Value.Should().BeAssignableTo<List<PickupLogDto>>().Subject;
-        history.Should().HaveCount(1);
-        history[0].CheckoutDateTime.Should().BeAfter(fromDate).And.BeBefore(toDate);
-    }
-
-    [Fact]
-    public async Task GetPickupHistory_WithNoHistory_ReturnsEmptyList()
-    {
-        // Arrange
-        var expectedHistory = new List<PickupLogDto>();
-
-        _authorizedPickupServiceMock
-            .Setup(s => s.GetPickupHistoryAsync(
-                _childIdKey,
-                null,
-                null,
-                It.IsAny<CancellationToken>()))
-            .ReturnsAsync(expectedHistory);
-
-        // Act
-        var result = await _supervisorController.GetPickupHistory(_childIdKey, null, null);
-
-        // Assert
-        var okResult = result.Should().BeOfType<OkObjectResult>().Subject;
-        var history = okResult.Value.Should().BeAssignableTo<List<PickupLogDto>>().Subject;
-        history.Should().BeEmpty();
     }
 
     #endregion
